[CX-RELEASE RELEVANT] chore: TRACEFOSS-2882 native query to criteria api replacement

+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/assets/asbuilt/infrastructure/repository/AssetAsBuiltRepositoryIT.java
@@ -45,7 +45,7 @@ class AssetAsBuiltRepositoryIT extends IntegrationTestSpecification {
     void givenIdField_whenGetFieldValues_thenSorted(
             String fieldName,
             String startWith,
-            Long resultLimit,
+            Integer resultLimit,
             Integer expectedSize
     ) {
         // given
@@ -62,11 +62,11 @@ class AssetAsBuiltRepositoryIT extends IntegrationTestSpecification {
 
     private static Stream<Arguments> fieldNameTestProvider() {
         return Stream.of(
-                Arguments.of("id", "urn:uuid:1", 10L, 3),
-                Arguments.of("id", null, 10L, 10),
-                Arguments.of("id", null, 200L, 13),
-                Arguments.of("inInvestigation", null, 10L, 1),
-                Arguments.of("owner", null, 10L, 2)
+                Arguments.of("id", "urn:uuid:1", 10, 3),
+                Arguments.of("id", null, 10, 10),
+                Arguments.of("id", null, 200, 13),
+                Arguments.of("inInvestigation", null, 10, 1),
+                Arguments.of("owner", null, 10, 2)
         );
     }
 }
